requirements.txt
django==4.0.5
python-environ==0.4.54
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-extra-fields==3.4.0
requests==2.28.0
faker==13.13.0
factory-boy==3.2.1